Add support for including the chunk timestamp in request parameters
Adds a new enable_chunk_time configuration boolean parameter. When set to true,
any endpoint_url with CHUNK_TIME_PLACEHOLDER will have that string literal
replaced with the timestamp at which the buffer chunk was created. This
information may be useful to upstream recipients of the HTTP request to gain
insight on the age of the data contained in the payload.